jupyter_tools: move template to separate file
